moved encryption to different file, added generating random key
